+++ hdt-lib/src/triples/TriplesKyoto.hpp
@@ -85,7 +85,7 @@ public:
 	 *
 	 * @return
 	 */
-	unsigned int getNumberOfElements();
+    size_t getNumberOfElements();
 
 	/**
 	 * Returns size in bytes
